add resolver

+++ src/main/java/br/com/higor/graphqlhibernate/repository/CommentRepository.java
@@ -9,5 +9,5 @@ import java.util.List;
 @Repository
 public interface CommentRepository extends JpaRepository<Comment, Integer> {
 
-    List<Comment> findAllByPostId(Integer id);
+    List<Comment> findAllByPostId(Long id);
 }
